% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K_big_lon_lat.tif}
\alias{K_big_lon_lat.tif}
\title{Example Of Carrying Capacity Map (Big)}
\format{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object with 74 rows
and 125 columns containing integer values 0-25 and NA's
indicating unsuitable areas.
}
\source{
Data generated in-house to serve as an example
(using spatial autocorrelation).
}
\description{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object representing a carrying
capacity map projected to WGS 84 (CRS84) from the original raster \code{K_big}.
This map can be used as a carrying capacity map to \code{\link{initialise}} data necessary
to perform a simulation with the \code{\link{sim}} function. It is compatible with the \code{n1_big_lon_lat.tif} raster.
}
\examples{
system.file("input_maps/K_big_lon_lat.tif", package = "rangr")

}
