% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{K_small.tif}
\alias{K_small.tif}
\title{Example of carrying capacity map (small)}
\format{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object with 15 rows
and 10 columns containing integer values 0-100 and NA's
(that indicates unsuitable areas).
}
\source{
Data generated in-house to serve as an example
(using spatial autocorrelation)
}
\description{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object that can be used as carrying
capacity map to \code{\link{initialise}} data necessary to perform a \code{\link{sim}}ulation.
This map is compatible with \code{\link{n1_small.tif}}
\itemize{
\item an \code{\link[terra:SpatRaster-class]{SpatRaster}} object that can be used
as simulation starting point.
}
}
\examples{
system.file("input_maps/K_small.tif", package = "rangr")
}
