% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K_small_changing.tif}
\alias{K_small_changing.tif}
\title{Example Of Changing Carrying Capacity Maps (Small)}
\format{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object with 3 layers,
each has 15 rows and 10 columns containing integer values 0-170 and
NA's that indicates unsuitable areas.
}
\source{
Data generated in-house to serve as an example
(using spatial autocorrelation).
}
\description{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object that can be used as carrying
capacity maps to \code{\link{initialise}} data necessary to perform a simulation
with the \code{\link{sim}} function.
To utilise these maps in \code{\link{initialise}} the user first must
use \code{\link{K_get_interpolation}} to generate a map for every time step
of the simulation.
These maps are compatible with \code{\link{n1_small.tif}}.
Each subsequent map contains a virtual environment with greater
carrying capacity than the previous one.
}
\examples{
system.file("input_maps/K_small_changing.tif", package = "rangr")

}
