% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{n1_big.tif}
\alias{n1_big.tif}
\title{Example of abundance map at first time step of the simulation (big)}
\format{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object with 100 rows
and 100 columns containing integer values 0-50 and NA's
(that indicates unsuitable areas).
}
\source{
Data generated in-house to serve as an example
}
\description{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object that can be used
as simulation starting point to \code{\link{initialise}} data necessary
to perform a \code{\link{sim}}ulation. This map is compatible with \code{\link{K_big.tif}} map.
}
\examples{
system.file("input_maps/n1_big.tif", package = "rangr")
}
