% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n1_big_lon_lat.tif}
\alias{n1_big_lon_lat.tif}
\title{Example Of Abundance Map At First Time Step Of The Simulation (Big)}
\format{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object with 74 rows
and 125 columns containing integer values 0-50 and NA's
indicating unsuitable areas.
}
\source{
Data generated in-house to serve as an example.
}
\description{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object representing an abundance map
at the first time step of the simulation projected to WGS 84 (CRS84) from the
original raster \code{n1_big}. This map can be used as a simulation starting point
to \code{\link{initialise}} data necessary to perform a simulation with the \code{\link{sim}}
function. It is compatible with the \code{K_big_lon_lat.tif} map.
}
\examples{
system.file("input_maps/n1_big_lon_lat.tif", package = "rangr")

}
