% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n1_small_lon_lat.tif}
\alias{n1_small_lon_lat.tif}
\title{Example Of Abundance Map At First Time Step Of The Simulation (Small)}
\format{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object with 12 rows
and 14 columns containing integer values 0-10 and NA's
(indicating unsuitable areas).
}
\source{
Data generated in-house to serve as an example
}
\description{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object representing an abundance map
at the first time step of the simulation projected to WGS 84 (CRS84) from the
original raster \code{n1_small}. This map can be used as a simulation starting point
to initialize data necessary to perform a simulation. It is compatible with the
\code{K_small_lon_lat.tif} and \code{K_small_changing_lon_lat.tif} maps.
}
\examples{
system.file("input_maps/n1_small_lon_lat.tif", package = "rangr")

}
