% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{observations_points}
\alias{observations_points}
\title{Example Of Observation Points List}
\format{
A data frame with 1500 rows and 3 variables:
\describe{
\item{x}{x coordinate}
\item{y}{y coordinate}
\item{time_step}{time_step at which the abundances should be observed}
}
}
\source{
Data generated in-house to serve as an example
}
\usage{
observations_points
}
\description{
A \code{data.frame} containing a sample input data to the function
\code{\link{get_observations}}  when \code{type} argument is set to "from_file".
This data is compatible with \code{\link{n1_small.tif}},
\code{\link{K_small.tif}} and \code{\link{K_small_changing.tif}} maps.
}
\keyword{datasets}
