% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangr-package.R
\docType{package}
\name{rangr-package}
\alias{rangr}
\alias{rangr-package}
\title{rangr: Mechanistic Simulation of Species Range Dynamics}
\description{
Integrates population dynamics and dispersal into a mechanistic virtual species simulator. The package can be used to study the effects of environmental change on population growth and range shifts. It allows for simple and straightforward definition of population dynamics (including positive density dependence), extensive possibilities for defining dispersal kernels, and the ability to generate virtual ecologist data. Learn more about the 'rangr' at \url{https://docs.ropensci.org/rangr/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/rangr}
  \item \url{https://docs.ropensci.org/rangr/}
  \item \doi{10.1111/2041-210X.14475}
  \item Report bugs at \url{https://github.com/ropensci/rangr/issues}
}

}
\author{
\strong{Maintainer}: Katarzyna Markowska \email{katarzyna.markowska@amu.edu.pl}

Authors:
\itemize{
  \item Katarzyna Markowska \email{katarzyna.markowska@amu.edu.pl}
  \item Lechosław Kuczyński \email{lechu@amu.edu.pl}
}

Other contributors:
\itemize{
  \item Tad Dallas [reviewer]
  \item Joanne Potts [reviewer]
}

}
\keyword{internal}
