% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{read_pos}
\alias{read_pos}
\title{Reads a HOMER peak/position file.}
\usage{
read_pos(file, keep_all = F, message = F)
}
\arguments{
\item{file}{A HOMER-formatted peak file (PeakID, chr, start, stop ...)}

\item{keep_all}{If \code{TRUE}, will keep all metadata columns.}

\item{message}{Include messages for column parsing?}
}
\value{
A GenomicRanges object
}
\description{
\code{read_pos} takes a HOMER peak file as input and tries to automatically find the columns that describe genomic ranges. It returns them as a GRanges object.
}
