% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_rast.R
\name{to_rast}
\alias{to_rast}
\title{Generic conversion to SpatRaster}
\usage{
to_rast(obj, ...)
}
\arguments{
\item{obj}{An object to convert.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{\link[terra:SpatRaster-class]{SpatRaster}} or a list of such objects.
}
\description{
A generic method to convert simulation result objects into
\code{\link[terra:SpatRaster-class]{SpatRaster}} format.
}
\examples{
\dontrun{
to_rast(sim_results_object)
}
}
\seealso{
\code{\link[=to_rast.sim_results]{to_rast.sim_results()}}
}
