% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{trim_GRanges}
\alias{trim_GRanges}
\title{Removes out-of-bounds intervals, given a specific genome build.}
\usage{
trim_GRanges(gr, genome = "hg19")
}
\arguments{
\item{gr}{a genomic ranges object.}

\item{genome}{a genome build. Defaults to "hg19"}
}
\value{
A genomic ranges object.
}
\description{
Removes out-of-bounds intervals, given a specific genome build.
}
