% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{truncate_cols}
\alias{truncate_cols}
\title{Truncates long column names matching pattern.}
\usage{
truncate_cols(df, pattern = pattern, replacement = replacement)
}
\arguments{
\item{df}{A data frame.}

\item{pattern}{character string containing a regular expression to trim.}

\item{replacement}{a replacement for matched pattern in sub and gsub.}
}
\value{
A dataframe with truncated columns.
}
\description{
Truncates long column names matching pattern.
}
\examples{
gr <- GRanges(Rle(c("chr2", "chr2", "chr1", "chr3"), c(1, 3, 2, 4)),
IRanges(1:10, width=10:1, names=head(letters, 10)),
Rle(strand(c("-", "+", "*", "+", "-")), c(1, 2, 2, 3, 2)),
score=1:10, GC=seq(1, 0, length=10))
df <- as.data.frame(gr)
df
df <- truncate_cols(df, pattern="re{0,}?.$")
df
}
