% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{write_mig}
\alias{write_mig}
\title{Save a GRanges or data frame object to tsv and UCSC-ready bed file for MIG.}
\usage{
write_mig(x, path = "", genome = "hg19", ucsc_suffix = ".ucsc.bed")
}
\arguments{
\item{x}{A \code{\link[GenomicRanges]{GenomicRanges}} or \code{data.frame} object.}

\item{path}{Path to write to.}

\item{genome}{A genome build, e.g. hg19}

\item{ucsc_suffix}{The suffix for the ucsc-ready bed file.}
}
\value{
writes an object to a specific path.
}
\description{
Save a GRanges or data frame object to tsv and UCSC-ready bed file for MIG.
}
\examples{
gr <- GRanges(Rle(c("chr2", "chr2", "chr1", "chr3"), c(1, 3, 2, 4)),
IRanges(1:10, width=10:1, names=head(letters, 10)),
Rle(strand(c("-", "+", "*", "+", "-")), c(1, 2, 2, 3, 2)),
score=1:10, GC=seq(1, 0, length=10))

write_bed(gr, path = "output/file.ucsc.bed", ucsc = TRUE)
}
