% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K_small.tif}
\alias{K_small.tif}
\title{Example Of Carrying Capacity Map (Small)}
\format{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object with 15 rows
and 10 columns containing integer values 0-100 and NA's indicating
unsuitable areas.
}
\source{
Data generated in-house to serve as an example
(using spatial autocorrelation).
}
\description{
\code{\link[terra:SpatRaster-class]{SpatRaster}} object that can be used a carrying
capacity map to \code{\link{initialise}} data necessary to perform a simulation
with the \code{\link{sim}} function.
This map is compatible with \code{\link{n1_small.tif}}.
}
\examples{
system.file("input_maps/K_small.tif", package = "rangr")

}
