% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangr-package.R
\docType{package}
\name{rangr-package}
\alias{rangr}
\alias{rangr-package}
\title{rangr: Mechanistic Simulation of Species Range Dynamics}
\description{
Set of tools for simulating species range dynamics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/popecol/rangr}
  \item \url{https://popecol.github.io/rangr/}
  \item Report bugs at \url{https://github.com/popecol/rangr/issues}
}

}
\author{
\strong{Maintainer}: Katarzyna Markowska \email{katarzyna.markowska@amu.edu.pl}

Authors:
\itemize{
  \item Lechosław Kuczyński \email{lechu@amu.edu.pl}
}

}
\keyword{internal}
