% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangr-package.R
\docType{package}
\name{rangr-package}
\alias{rangr}
\alias{rangr-package}
\title{rangr: Mechanistic Simulation of Species Range Dynamics}
\description{
Species range dynamics simulation toolset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/rangr}
  \item \url{https://docs.ropensci.org/rangr/}
  \item Report bugs at \url{https://github.com/ropensci/rangr/issues}
}

}
\author{
\strong{Maintainer}: Katarzyna Markowska \email{katarzyna.markowska@amu.edu.pl}

Authors:
\itemize{
  \item Lechosław Kuczyński \email{lechu@amu.edu.pl}
}

Other contributors:
\itemize{
  \item Tad Dallas (@taddallas) [reviewer]
  \item Joanne Potts (@TheAnalyticalEdge) [reviewer]
}

}
\keyword{internal}
