% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{trim_GRanges}
\alias{trim_GRanges}
\title{Remove out-of-bounds ranges, using a given genome build}
\usage{
trim_GRanges(gr, genome = "")
}
\arguments{
\item{gr}{A \code{GenomicRanges} object.}

\item{genome}{A genome build, e.g. hg19}
}
\value{
a filtered \code{GenomicRanges} object.
}
\description{
Remove out-of-bounds ranges, using a given genome build
}
